Add Dropbox Paper convenience method for making throwaway Paper files
This is run at the beginning of tests for the get_metadata,
get_folder_info, and download endpoints so that a user running the
test doesn't have to set an environment variable to the ID of one of
their Papers, outside of the TestPaper_GetFolderInfoInsideFolders test,
since there isn't an easy way to make a throwaway file that is inside
a folder.

For these tests that create throwaway files, the environment variable
TEST_CAN_CREATE_DBOX_PAPER must be set so a user knows that they are
running tests that do writes to their Dropbox.